Re-factor RPCs in two ways.
First, the distinction between an RPC layer, which marshalled
arguments in and out of lower-level "request" buffers, and a request
layer, which sent a request buffer and matched it up with a
subsequent response, is now a single layer. This avoids having to
define two structures for each RPC (in and out) as well as all of
their copying.

Second, the use of a large C union to define each RPC request type is
now eschewed in favor of a serialization/deserialization layer.
Callers now pack their arguments and unpack the responses.

These changes greatly streamline how easily a new RPC can be added.
This is in preparation many more new RPC's to come soon.